refactor: useUser 훅 분리
