+++ src/components/hoc/withPrivateRoute.tsx
@@ -1,6 +1,7 @@
 import { ALERT_MESSAGE } from 'constants/index';
 import { useAppDispatch } from 'hooks/useAppDispatch';
 import { useAppSelector } from 'hooks/useAppSelector';
+import useUser from 'hooks/useUser';
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { getUser } from 'redux/user/thunk';
@@ -8,21 +9,18 @@ import { PATH } from 'Routers';
 
 const withPrivateRoute = (Component: React.ComponentType<unknown>) => {
   return function Wrapper({ ...props }) {
-    const isLogin = useAppSelector(state => !!state.user.data);
-    const dispatch = useAppDispatch();
     const navigate = useNavigate();
 
-    useEffect(() => {
-      const accessToken = localStorage.getItem('access-token');
-
-      if (!accessToken) return;
-
-      dispatch(getUser()).catch(() => {
+    const { isLogin } = useUser({
+      onFailure: () => {
         alert(ALERT_MESSAGE.WRONG_ACCESS);
-        localStorage.removeItem('access-token');
         navigate(PATH.login);
-      });
-    }, []);
+      },
+      withoutToken: () => {
+        alert(ALERT_MESSAGE.WRONG_ACCESS);
+        navigate(PATH.login);
+      },
+    });
 
     if (!isLogin) return null;
