+++ src/components/hoc/withPublicRoute.tsx
@@ -1,31 +1,19 @@
 import { ALERT_MESSAGE } from 'constants/index';
-import { useAppDispatch } from 'hooks/useAppDispatch';
-import { useAppSelector } from 'hooks/useAppSelector';
-import React, { useEffect } from 'react';
+import useUser from 'hooks/useUser';
+import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import { getUser } from 'redux/user/thunk';
 import { PATH } from 'Routers';
 
 const withPublicRoute = (Component: React.ComponentType<unknown>) => {
   return function Wrapper({ ...props }) {
-    const isLogin = useAppSelector(state => !!state.user.data);
-    const dispatch = useAppDispatch();
     const navigate = useNavigate();
 
-    useEffect(() => {
-      const accessToken = localStorage.getItem('access-token');
-
-      if (!accessToken) return;
-
-      dispatch(getUser())
-        .then(() => {
-          alert(ALERT_MESSAGE.WRONG_ACCESS);
-          navigate(PATH.home);
-        })
-        .catch(() => {
-          localStorage.removeItem('access-token');
-        });
-    }, []);
+    const { isLogin } = useUser({
+      onSuccess: () => {
+        alert(ALERT_MESSAGE.WRONG_ACCESS);
+        navigate(PATH.home);
+      },
+    });
 
     if (isLogin) return null;
